backend/app/requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
cachelib==0.10.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
cryptography==39.0.0
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.6
Fiona==1.8.22
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Session==0.4.0
firebase==4.0.1
future==0.18.2
gcloud==0.18.3
GeoAlchemy2==0.12.5
geocoder==1.38.1
geopandas==0.12.2
googleapis-common-protos==1.58.0
googlemaps==4.7.3
greenlet @ file:///C:/b/abs_47lk_w2ajq/croot/greenlet_1670513248400/work
httplib2==0.21.0
idna==3.4
image==1.5.33
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsons==1.6.3
jwcrypto==1.4.2
MarkupSafe==2.1.1
munch==2.5.0
neo4j==5.3.0
numpy==1.24.1
oauth2client==4.1.3
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
protobuf==4.22.0rc2
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pyparsing==3.0.9
pyproj==3.4.1
Pyrebase4==4.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jwt==4.0.0
pytz==2022.7
ratelim==0.1.6
regex==2022.10.31
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
shapely==2.0.0
six==1.16.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
typish==1.9.3
tzdata==2022.7
urllib3==1.26.13
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.11.0
